addAWS(2)
